Removal of the IntegrationTest scope usage from build.sbt
