+++ .github/workflows/pr.yml
@@ -253,52 +253,52 @@ jobs:
       - name: Compile
         shell: bash
         run: CROSS_BUILD=true ./ciRunSbt.sh Compile/compile Test/compile
-  integrationTests:
-    name: IntegrationTests
-    needs: [ build, setup ]
-    strategy:
-      matrix:
-        scalaVersion: ${{fromJson(needs.setup.outputs.scala_version_matrix)}}
-    env:
-      SCALA_VERSION: ${{ matrix.scalaVersion }}
-    if: ${{ needs.setup.outputs.fe_changes_count != needs.setup.outputs.all_changes_count || github.ref == 'refs/heads/staging' || github.ref == 'refs/heads/master' || github.ref == 'refs/heads/demo' || startsWith(github.ref, 'refs/heads/preview') || startsWith(github.ref, 'refs/heads/release') }}
-    runs-on: ubuntu-latest
-    steps:
-      - name: Cancel previous runs
-        if: github.event_name != 'push'
-        uses: styfle/cancel-workflow-action@0.11.0
-        with:
-          access_token: ${{ secrets.GITHUB_TOKEN }}
-      - uses: actions/checkout@v3
-      - uses: coursier/setup-action@v1
-        with:
-          jvm: temurin:1.11.0.17
-      - name: Cache ivy packages
-        uses: actions/cache@v3
-        with:
-          path: |
-            ~/.ivy2/cache
-            ~/.cache/coursier
-            ~/.sbt
-          key: sbt-cache-${{ hashFiles('**/*.sbt') }}
-          restore-keys: sbt
-      - uses: actions/download-artifact@v3
-        with:
-          name: build-target
-      - name: Untar artifacts
-        shell: bash
-        run:  tar xfz target.tgz
-      - name: Integration tests
-        shell: bash
-        env:
-          dockerUpLatest: true
-        run: ./ciRunSbt.sh It/test
-      - name: Test Report
-        if: success() || failure()
-        uses: actions/upload-artifact@v3
-        with:
-          name: IntegrationTests-results-${{ matrix.scalaVersion }}
-          path: '**/it-reports/*.xml'
+#  integrationTests:
+#    name: IntegrationTests
+#    needs: [ build, setup ]
+#    strategy:
+#      matrix:
+#        scalaVersion: ${{fromJson(needs.setup.outputs.scala_version_matrix)}}
+#    env:
+#      SCALA_VERSION: ${{ matrix.scalaVersion }}
+#    if: ${{ needs.setup.outputs.fe_changes_count != needs.setup.outputs.all_changes_count || github.ref == 'refs/heads/staging' || github.ref == 'refs/heads/master' || github.ref == 'refs/heads/demo' || startsWith(github.ref, 'refs/heads/preview') || startsWith(github.ref, 'refs/heads/release') }}
+#    runs-on: ubuntu-latest
+#    steps:
+#      - name: Cancel previous runs
+#        if: github.event_name != 'push'
+#        uses: styfle/cancel-workflow-action@0.11.0
+#        with:
+#          access_token: ${{ secrets.GITHUB_TOKEN }}
+#      - uses: actions/checkout@v3
+#      - uses: coursier/setup-action@v1
+#        with:
+#          jvm: temurin:1.11.0.17
+#      - name: Cache ivy packages
+#        uses: actions/cache@v3
+#        with:
+#          path: |
+#            ~/.ivy2/cache
+#            ~/.cache/coursier
+#            ~/.sbt
+#          key: sbt-cache-${{ hashFiles('**/*.sbt') }}
+#          restore-keys: sbt
+#      - uses: actions/download-artifact@v3
+#        with:
+#          name: build-target
+#      - name: Untar artifacts
+#        shell: bash
+#        run:  tar xfz target.tgz
+#      - name: Integration tests
+#        shell: bash
+#        env:
+#          dockerUpLatest: true
+#        run: ./ciRunSbt.sh It/test
+#      - name: Test Report
+#        if: success() || failure()
+#        uses: actions/upload-artifact@v3
+#        with:
+#          name: IntegrationTests-results-${{ matrix.scalaVersion }}
+#          path: '**/it-reports/*.xml'
 
   slowTests:
     name: SlowTests
@@ -352,7 +352,9 @@ jobs:
           AZURE_TENANT_ID: ${{ secrets.AZURE_TENANT_ID }}
         shell: bash
 #        run: ./ciRunSbt.sh designer/Slow/test liteK8sDeploymentManager/ExternalDepsTests/test schemedKafkaComponentsUtils/ExternalDepsTests/test liteKafkaComponentsTests/ExternalDepsTests/test
-        run: ./ciRunSbt.sh designer/Slow/test schemedKafkaComponentsUtils/ExternalDepsTests/test liteKafkaComponentsTests/ExternalDepsTests/test
+        run: ./ciRunSbt.sh security/Slow/test processReports/Slow/test openapiComponents/Slow/test flinkDeploymentManager/Slow/test 
+          kafkaComponentsUtils/Slow/test liteEngineKafkaIntegrationTest/Slow/test security/Slow/test openapiComponents/Slow/test
+          designer/Slow/test schemedKafkaComponentsUtils/ExternalDepsTests/test liteKafkaComponentsTests/ExternalDepsTests/test
 #      - name: docker logs
 #        if: success() || failure()
 #        run: |
@@ -566,7 +568,7 @@ jobs:
   #TODO: extract to different workflow?
   publish:
     runs-on: ubuntu-latest
-    needs: [build, build-fe, setup, tests, crossCompile, integrationTests, slowTests, frontendTests, cypressTests]
+    needs: [build, build-fe, setup, tests, crossCompile, slowTests, frontendTests, cypressTests]
     #TODO: should we publish everything on all those branches?
     if: ${{ github.ref == 'refs/heads/staging' || github.ref == 'refs/heads/master' || github.ref == 'refs/heads/demo' || startsWith(github.ref, 'refs/heads/preview') || startsWith(github.ref, 'refs/heads/release') }}
     env:
