+++ components/openapi/src/test/resources/application.conf
