+++ components/openapi/src/test/resources/customer-primitive-return-swagger.yaml
