+++ components/openapi/src/test/resources/customer-primitive-swagger.yaml
