+++ components/openapi/src/test/resources/customer-swagger.json
