+++ components/openapi/src/test/scala/pl/touk/nussknacker/openapi/functional/StubService.scala
