+++ engine/flink/management/src/test/resources/docker/entrypointWithIP.sh
