+++ engine/lite/integration-test/src/test/scala/pl/touk/nussknacker/engine/lite/kafka/NuKafkaRuntimeTestMixin.scala
