+++ engine/lite/integration-test/src/test/scala/pl/touk/nussknacker/engine/lite/kafka/sample/NuKafkaRuntimeTestSamples.scala
