+++ engine/lite/integration-test/src/test/scala/pl/touk/nussknacker/engine/lite/utils/BaseNuRuntimeBinTestMixin.scala
