+++ security/src/test/resources/keycloak.json
